Refactor omp transcript parsing
Separate session decoding from transcript analysis and give full sessions and checkpoint-scoped slices explicit parsing paths. Keep header validation strict whenever a title or session header is present, and document and test the accepted transcript shapes.
